+++ eeos/src/main/java/com/blackcompany/eeos/comment/presentation/controller/CommentController.java
@@ -13,10 +13,10 @@ import com.blackcompany.eeos.comment.application.usecase.DeleteCommentUsecase;
 import com.blackcompany.eeos.comment.application.usecase.GetCommentUsecase;
 import com.blackcompany.eeos.comment.application.usecase.UpdateCommentUsecase;
 import com.blackcompany.eeos.comment.presentation.docs.CommentApi;
-import com.blackcompany.eeos.common.presentation.respnose.ApiResponse;
-import com.blackcompany.eeos.common.presentation.respnose.ApiResponseBody.SuccessBody;
-import com.blackcompany.eeos.common.presentation.respnose.ApiResponseGenerator;
-import com.blackcompany.eeos.common.presentation.respnose.MessageCode;
+import com.blackcompany.eeos.common.presentation.response.ApiResponse;
+import com.blackcompany.eeos.common.presentation.response.ApiResponseBody.SuccessBody;
+import com.blackcompany.eeos.common.presentation.response.ApiResponseGenerator;
+import com.blackcompany.eeos.common.presentation.response.MessageCode;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
